ray header file completed

+++ color.h
@@ -1,15 +1,16 @@
 #ifndef COLOR_H
 #define COLOR_H
 
-#include <iostream>
 #include "vec3.h"
+#include <iostream>
 
 using color = vec3;
 
-void write_color (std::ostream& out, const color& pixel_color){
-    //Each Pixel has a value between 0 and 1 so scale by 255.9999 to get the true color
-    out << (255.999 * pixel_color.x()) << ' ' << (255.999 * pixel_color.y()) << ' ' << (255.999 * pixel_color.z()) << '\n';
+void write_color(std::ostream& out, const color& pixel_color) {
+    // Write the translated [0,255] value of each color component.
+    out << int(255.999 * pixel_color.x()) << ' '
+        << int(255.999 * pixel_color.y()) << ' '
+        << int(255.999 * pixel_color.z()) << '\n';
 }
-
 
 #endif
